+++ custom_components/legrand_energy/models/__init__.py
@@ -8,9 +8,11 @@ from .energy import (
     LegrandProjections,
 )
 from .fluid import FluidType
+from .fluid_measurements import FluidMeasurements
 
 __all__ = [
     "FluidType",
+    "FluidMeasurements",
     "LegrandEnergyData",
     "LegrandMeasurements",
     "LegrandModule",